requirements.txt
Flask==2.0.3
gunicorn
google-cloud-speech==2.7.0
google-cloud-language==2.10.0
